pyproject.toml
[tool.poetry]
name = "covid-col-map"
version = "0.1.0"
description = ""
authors = ["charlielito"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.3"
typer = "^0.1.1"
plotly = "^4.6.0"
unidecode = "^1.1.1"
geopy = "^1.21.0"
sodapy = "^2.1.0"
boto3 = "^1.12.39"
s3fs = "^0.4.2"
python-dateutil = "^2.8.1"
psutil = "^5.7.0"
Pillow = "^7.1.2"
tqdm = "^4.46.0"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
